Move the tests for AbstractConv with the implementation.
Also makes sure that the CPU versions are tested even if there are no GPUs.
